optimize ETL forward compatibility

+++ src/SmartCode.ETL/IExtractData.cs
@@ -4,7 +4,7 @@ namespace SmartCode.ETL
 {
     public interface IExtractData : IPlugin
     {
-        long Total { get; }
+        int Total { get; }
         int BulkSize { get; }
         int Offset { get; }
         Task Run();
